move bone categories to service
